I was investigating https://github.com/dinoki-ai/osaurus/issues/47, and have had success in optimizing swift-transformers, and noticed that there were few optimizations to be made in Jinja.

Key Changes:

- `Template.render` was recreating `Environment` each time. Created sharable environment
- Small algorithm changes, i.e. remove string concat in favor or arrays

Benchmark Results:

| Phase              | Before (ms) | After (ms) | Change | % Improvement |
|--------------------|-------------|------------|--------|---------------|
| **Tokenize avg**   | 69.525      | 68.887     | -0.638 | ~0.9% faster  |
| **Parse avg**      | 0.498       | 0.457      | -0.041 | ~8.2% faster  |
| **Run avg**        | 0.116       | 0.099      | -0.017 | ~14.7% faster |
| **Template.render**| 0.176       | 0.142      | -0.034 | ~19.3% faster |