@johnmai-dev 2025-09-06
Looks great, thank you! ♥️

Could you fix the format CI failure? 

```bash
swift-format . -i -r
```