@tpae 2025-09-07
> > Thank you for the edits and comments. I can confirm the new optimizations are working.
> > Phase	Before (ms)	After (ms)	Change	% Improvement
> > **Tokenize avg**	70.770	69.055	-1.715	~2.4% faster
> > **Parse avg**	0.492	0.467	-0.025	~5.1% faster
> > **Run avg**	0.113	0.101	-0.012	~10.6% faster
> > **Template.render**	0.173	0.143	-0.030	~17.3% faster
> 
> Looks like the improvement dropped a little bit.

It hasn't dropped, the benchmarks are slightly different each run. It depends on the system at that given time, I think.